Begin adding DHCP
git-svn-id: https://svn.code.sf.net/p/netemul/svn/trunk@113 8abb0c59-4564-46b7-95fe-082762c05281

+++ src/programms/ripprogramm.cpp
@@ -189,7 +189,7 @@ void ripProgramm::addToTemp(routeRecord *r)
 */
 void ripProgramm::write(QDataStream &stream) const
 {
-    stream << (int)RIP;
+    stream << RIP;
     programmRep::write(stream);
 }
 //---------------------------------------------------
